complete Assignment1 and start 2

+++ Assignement1/CalculationExamples/src/calculationexample/CalculationExample.java
@@ -11,16 +11,16 @@ public class CalculationExample {
         Scanner scanner = new Scanner(System.in);
 
         System.out.print("Enter a number to calculate its factorial: ");
-        Integer num = Integer.valueOf(scanner.nextInt());
+        int num = scanner.nextInt();
         System.out.println("Factorial of " + num + " is: " + calc.factorial(num));
 
         System.out.print("Enter a number to check if it is prime: ");
-        int primeCheck = scanner.nextInt();
+        long primeCheck = scanner.nextLong();
         System.out.println(primeCheck + " is prime: " + calc.isPrime(primeCheck));
 
         System.out.print("Enter two numbers to find their GCD: ");
-        int num1 = scanner.nextInt();
-        int num2 = scanner.nextInt();
+        long num1 = scanner.nextLong();
+        long num2 = scanner.nextLong();
         System.out.println("GCD of " + num1 + " and " + num2 + " is: " + calc.gcd(num1, num2));
 
         scanner.close();
